Minor tweaks and `make check` passes.
