+++ TypeSystem/optional.h
@@ -39,7 +39,7 @@ SOFTWARE.
 
 #include <memory>
 
-#include "sfinae.h"
+#include "types.h"
 #include "exceptions.h"
 
 #include "../Bricks/template/enable_if.h"
